feat: util 이름 변경 및 로직 분리
